feat: add RESERVATION_COLORS and move RESERVATION_LABELS to types
