+++ src/types/reservation.ts
@@ -33,3 +33,19 @@ export interface Reservation {
   createdAt: string;
   updatedAt: string;
 }
+
+export const RESERVATION_LABELS: Record<ReservationStatus, string> = {
+  pending: '예약 신청',
+  confirmed: '예약 승인',
+  declined: '예약 거절',
+  canceled: '예약 취소',
+  completed: '체험 완료',
+};
+
+export const RESERVATION_COLORS: Record<ReservationStatus, string> = {
+  pending: 'text-var-blue-dark',
+  confirmed: 'text-var-orange-dark',
+  declined: 'text-var-red-dark',
+  canceled: 'text-var-gray2',
+  completed: 'text-var-gray2',
+};
